+++ webpack-server.js
@@ -8,8 +8,10 @@ const args = require('./build-arguments');
 
 const webpack = require('webpack');
 const webpackMerge = require('webpack-merge');
-const WebpackDevServer = require('webpack-dev-server');
+const middleware = require('webpack-dev-middleware');
+const hotMiddleware = require('webpack-hot-middleware');
 const getWebpackConfig = require('./get-webpack-config');
+const express = require('express');
 
 if (!args.config) {
     throw new Error('Please provide config path --config=PATH_TO_CONFIG.js');
@@ -47,7 +49,6 @@ const tsConfigPath = config.tsConfigPath
     : path.resolve(process.cwd(), './tsconfig.json');
 
 const ourWebpackConfig = getWebpackConfig({
-    devServerUrl: `http://${HOST}:${PORT}`,
     buildDir: args.buildDir,
     configPath,
     getSections: config.getSections,
@@ -83,8 +84,16 @@ if (isCompiling) {
         console.log('Styleguide compiled');
     });
 } else {
-    const devServerOptions = Object.assign({}, ourWebpackConfig.devServer);
-    const server = new WebpackDevServer(compiler, devServerOptions);
+    const server = express();
+
+    server
+        .use(
+            middleware(compiler, {
+                publicPath: mergedConfig.output.publicPath,
+            })
+        )
+        .use(hotMiddleware(compiler));
+
     server.listen(PORT, HOST, () => {
         console.log(`Starting server on http://${HOST}:${PORT}`);
     });